Move workspace abstract types to db/meta
