+++ packages/db/src/loaders/commands/names.ts
@@ -4,12 +4,8 @@ import {
 } from "@truffle/db/loaders/resources/projects";
 
 import { generateNameRecordsLoad } from "@truffle/db/loaders/resources/nameRecords";
-import {
-  WorkspaceRequest,
-  WorkspaceResponse,
-  IdObject,
-  NamedResource
-} from "@truffle/db/loaders/types";
+import { WorkspaceRequest, WorkspaceResponse } from "@truffle/db/loaders/types";
+import { IdObject, NamedResource } from "@truffle/db/meta";
 
 /**
  * generator function to load nameRecords and project names into Truffle DB
