+++ packages/db/src/loaders/resources/bytecodes/index.ts
@@ -1,10 +1,10 @@
 import {
   CompilationData,
-  toIdObject,
   LoadedBytecodes,
   WorkspaceRequest,
   WorkspaceResponse
 } from "@truffle/db/loaders/types";
+import { toIdObject } from "@truffle/db/meta";
 import { CompiledContract } from "@truffle/compile-common";
 import { AddBytecodes } from "./add.graphql";
 export { AddBytecodes };
