+++ packages/db/src/loaders/resources/sources/index.ts
@@ -1,7 +1,8 @@
+import { toIdObject } from "@truffle/db/meta";
+
 import {
   CompilationData,
   LoadedSources,
-  toIdObject,
   WorkspaceRequest,
   WorkspaceResponse
 } from "@truffle/db/loaders/types";
