+++ packages/db/src/workspace/pouch/memory.ts
@@ -1,8 +1,8 @@
 import PouchDB from "pouchdb";
 import PouchDBMemoryAdapter from "pouchdb-adapter-memory";
 
+import { Collections } from "@truffle/db/meta";
 import { Databases } from "./databases";
-import { Collections } from "./types";
 
 export class MemoryDatabases<C extends Collections> extends Databases<C> {
   static counter: number = 0;
