+++ packages/db/src/workspace/pouch/sqlite.ts
@@ -3,8 +3,8 @@ import fse from "fs-extra";
 import PouchDB from "pouchdb";
 import PouchDBNodeWebSQLAdapter from "pouchdb-adapter-node-websql";
 
+import { Collections } from "@truffle/db/meta";
 import { Databases } from "./databases";
-import { Collections } from "./types";
 
 export class SqliteDatabases<C extends Collections> extends Databases<C> {
   private directory: string;
